# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Việt phục Remix API",
    "description": "Backend REST API phục vụ nền tảng phối đồ Việt phục Remix (VietStylist)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/catalog/items": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List Items",
        "description": "Lấy danh sách món đồ trang phục, quần và phụ kiện với bộ lọc phong phú.",
        "operationId": "list_items_api_catalog_items_get",
        "parameters": [
          {
            "name": "garment_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lọc theo nhóm áo",
              "title": "Garment Type Id"
            },
            "description": "Lọc theo nhóm áo"
          },
          {
            "name": "slot",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lọc theo vị trí (outerwear, bottom, headwear...)",
              "title": "Slot"
            },
            "description": "Lọc theo vị trí (outerwear, bottom, headwear...)"
          },
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lọc theo giới tính (male, female, unisex)",
              "title": "Gender"
            },
            "description": "Lọc theo giới tính (male, female, unisex)"
          },
          {
            "name": "occasion_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lọc theo sự kiện",
              "title": "Occasion Id"
            },
            "description": "Lọc theo sự kiện"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tìm kiếm tên hoặc mô tả",
              "title": "Search"
            },
            "description": "Tìm kiếm tên hoặc mô tả"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemSummaryResponse"
                  },
                  "title": "Response List Items Api Catalog Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/items/{item_id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Get Item Detail",
        "description": "Xem chi tiết một món đồ gồm các biến thể màu sắc, lớp ảnh vector 2D và sự kiện phù hợp.",
        "operationId": "get_item_detail_api_catalog_items__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lookbooks": {
      "get": {
        "tags": [
          "Lookbooks"
        ],
        "summary": "List Lookbooks",
        "description": "Lấy danh sách lookbooks cá nhân của người dùng đã đăng nhập.",
        "operationId": "list_lookbooks_api_lookbooks_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LookbookResponse"
                  },
                  "title": "Response List Lookbooks Api Lookbooks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lookbooks"
        ],
        "summary": "Create Lookbook",
        "description": "Tạo mới một bộ sưu tập Lookbook (F09).",
        "operationId": "create_lookbook_api_lookbooks_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLookbookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookbookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lookbooks/{lookbook_id}": {
      "get": {
        "tags": [
          "Lookbooks"
        ],
        "summary": "Get Lookbook",
        "description": "Xem chi tiết lookbook và danh sách các bộ phối đính kèm.",
        "operationId": "get_lookbook_api_lookbooks__lookbook_id__get",
        "parameters": [
          {
            "name": "lookbook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lookbook Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookbookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Lookbooks"
        ],
        "summary": "Update Lookbook",
        "description": "Cập nhật tiêu đề, mô tả hoặc danh sách bộ phối trong lookbook.",
        "operationId": "update_lookbook_api_lookbooks__lookbook_id__put",
        "parameters": [
          {
            "name": "lookbook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lookbook Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLookbookRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookbookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Lookbooks"
        ],
        "summary": "Delete Lookbook",
        "description": "Xóa lookbook.",
        "operationId": "delete_lookbook_api_lookbooks__lookbook_id__delete",
        "parameters": [
          {
            "name": "lookbook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lookbook Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lookbooks/{lookbook_id}/share": {
      "post": {
        "tags": [
          "Lookbooks"
        ],
        "summary": "Share Lookbook",
        "description": "Tạo liên kết chia sẻ công khai hoặc unlisted với token băm ngẫu nhiên bảo mật cao (F09).",
        "operationId": "share_lookbook_api_lookbooks__lookbook_id__share_post",
        "parameters": [
          {
            "name": "lookbook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lookbook Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareLinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetLayerResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "variant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          },
          "avatar_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Id"
          },
          "slot": {
            "type": "string",
            "title": "Slot"
          },
          "z_index": {
            "type": "integer",
            "title": "Z Index",
            "default": 10
          },
          "anchor_x": {
            "type": "number",
            "title": "Anchor X",
            "default": 0.0
          },
          "anchor_y": {
            "type": "number",
            "title": "Anchor Y",
            "default": 0.0
          },
          "scale_x": {
            "type": "number",
            "title": "Scale X",
            "default": 1.0
          },
          "scale_y": {
            "type": "number",
            "title": "Scale Y",
            "default": 1.0
          },
          "layer_type": {
            "type": "string",
            "title": "Layer Type",
            "default": "svg"
          },
          "svg_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Svg Content"
          },
          "media_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Asset Id"
          },
          "color_mask_rule": {
            "type": "object",
            "title": "Color Mask Rule",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "slot"
        ],
        "title": "AssetLayerResponse"
      },
...
      "CreateLookbookRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "default": "Lookbook Cổ Phục"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility",
            "default": "private"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LookbookEntryInput"
            },
            "type": "array",
            "title": "Entries",
            "default": []
          }
        },
        "type": "object",
        "title": "CreateLookbookRequest"
      },
...
      "CreateShareLinkRequest": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": "view_only"
          },
          "expires_in_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Days",
            "default": 30
          }
        },
        "type": "object",
        "title": "CreateShareLinkRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "garment_type_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Garment Type Id"
          },
          "slot": {
            "type": "string",
            "title": "Slot"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender": {
            "type": "string",
            "title": "Gender",
            "default": "unisex"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "era": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Era"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "default": true
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "default": {}
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ItemVariantResponse"
            },
            "type": "array",
            "title": "Variants",
            "default": []
          },
          "default_layer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetLayerResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "asset_layers": {
            "items": {
              "$ref": "#/components/schemas/AssetLayerResponse"
            },
            "type": "array",
            "title": "Asset Layers",
            "default": []
          },
          "occasions": {
            "items": {
              "$ref": "#/components/schemas/OccasionResponse"
            },
            "type": "array",
            "title": "Occasions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "slot",
          "name"
        ],
        "title": "ItemDetailResponse"
      },
      "ItemSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "garment_type_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Garment Type Id"
          },
          "slot": {
            "type": "string",
            "title": "Slot"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender": {
            "type": "string",
            "title": "Gender",
            "default": "unisex"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "era": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Era"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "default": true
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "default": {}
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ItemVariantResponse"
            },
            "type": "array",
            "title": "Variants",
            "default": []
          },
          "default_layer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetLayerResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "slot",
          "name"
        ],
        "title": "ItemSummaryResponse"
      },
...
      "ItemVariantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "color_name": {
            "type": "string",
            "title": "Color Name"
          },
          "hex_color": {
            "type": "string",
            "title": "Hex Color"
          },
          "secondary_hex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secondary Hex"
          },
          "material": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material",
            "default": "Lụa tơ tằm"
          },
          "thickness_level": {
            "type": "string",
            "title": "Thickness Level",
            "default": "medium"
          },
          "pattern_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern Description"
          },
          "price_tier": {
            "type": "string",
            "title": "Price Tier",
            "default": "standard"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "color_name",
          "hex_color"
        ],
        "title": "ItemVariantResponse"
      },
...
      "LookbookResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LookbookEntryDetail"
            },
            "type": "array",
            "title": "Entries",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "title",
          "visibility",
          "created_at",
          "updated_at"
        ],
        "title": "LookbookResponse"
      },
...
      "ShareLinkResponse": {
        "properties": {
          "share_token": {
            "type": "string",
            "title": "Share Token"
          },
          "share_url": {
            "type": "string",
            "title": "Share Url"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "share_token",
          "share_url",
          "scope"
        ],
        "title": "ShareLinkResponse"
      },
...
      "UpdateLookbookRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          },
          "entries": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LookbookEntryInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entries"
          }
        },
        "type": "object",
        "title": "UpdateLookbookRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}